Skip unsupported tests instead of xfail
When a simulator does not support a test, skip it instead of
xfailing. This does not actually change how the test is run,
because the way we were using xfail it would not actually
execute the test, so skip will do the same. Now tests will
be marked as skipped in the log, to differentiate them from
tests that are marked xfail and run (which could potentially xpass).